+++ tests/test_deployment.py
@@ -5,6 +5,7 @@ import mock
 import os
 import pexpect
 import pytest
+import re
 import six
 import sys
 import tempfile
@@ -736,7 +737,9 @@ class TestDebugServer(object):
             __enter__=mock.Mock(side_effect=OSError),
             __exit__=mock.Mock(return_value=False),
         )
-        with mock.patch("dallinger.deployment.HerokuLocalWrapper") as Wrapper:
+        with mock.patch(
+            "dallinger.deployment.HerokuLocalDeployment.WRAPPER_CLASS"
+        ) as Wrapper:
             Wrapper.return_value = mock_wrapper
             with pytest.raises(OSError):
                 debugger.run()
@@ -814,7 +817,7 @@ class TestDebugServer(object):
     def test_failure(self, debugger):
         from requests.exceptions import HTTPError
 
-        with mock.patch("dallinger.deployment.HerokuLocalWrapper"):
+        with mock.patch("dallinger.deployment.HerokuLocalDeployment.WRAPPER_CLASS"):
             with mock.patch("dallinger.deployment.requests.post") as mock_post:
                 mock_post.return_value = mock.Mock(
                     ok=False,
@@ -835,6 +838,67 @@ class TestDebugServer(object):
                 mock.call("msg!"),
             ]
         )
+
+
+@pytest.mark.usefixtures("bartlett_dir", "clear_workers", "env")
+@pytest.mark.slow
+@pytest.mark.docker
+class TestDockerServer(object):
+    @pytest.mark.usefixtures("check_runbot")
+    def test_docker_debug_with_bots(self, env):
+        # Make sure debug server runs to completion with bots
+        p = pexpect.spawn(
+            "dallinger",
+            ["docker", "debug", "--verbose", "--bot", "--no-browsers"],
+            env=env,
+            encoding="utf-8",
+        )
+        p.logfile = sys.stdout
+        try:
+            p.expect_exact("Server is running", timeout=300)
+            p.expect_exact("Recruitment is complete", timeout=600)
+            p.expect_exact("'status': 'success'", timeout=60)
+            p.expect_exact("Experiment completed", timeout=10)
+            p.expect_exact("Stopping bartlett1932_web_1", timeout=20)
+            p.expect(pexpect.EOF)
+        finally:
+            try:
+                p.sendcontrol("c")
+                p.read()
+            except IOError:
+                pass
+
+    @pytest.mark.usefixtures("check_runbot")
+    def test_docker_debug_without_bots(self, env):
+        sys.path.append(os.getcwd())
+        from experiment import Bot
+
+        # Make sure debug server runs to completion without bots
+        p = pexpect.spawn(
+            "dallinger",
+            ["docker", "debug", "--verbose", "--no-browsers"],
+            env=env,
+            encoding="utf-8",
+        )
+        p.logfile = sys.stdout
+        try:
+            p.expect_exact("Server is running", timeout=60)
+            p.expect_exact("Initial recruitment list:", timeout=20)
+            p.expect("New participant requested.*", 30)
+            Bot(re.search("http://[^ \n\r]+", p.after).group()).run_experiment()
+            p.expect("New participant requested.*", 30)
+            Bot(re.search("http://[^ \n\r]+", p.after).group()).run_experiment()
+            p.expect_exact("Recruitment is complete", timeout=60)
+            p.expect_exact("'status': 'success'", timeout=60)
+            p.expect_exact("Experiment completed", timeout=10)
+            p.expect_exact("Stopping bartlett1932_web_1", timeout=20)
+            p.expect(pexpect.EOF)
+        finally:
+            try:
+                p.sendcontrol("c")
+                p.read()
+            except IOError:
+                pass
 
 
 @pytest.mark.usefixtures("bartlett_dir", "clear_workers", "env")