+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 ## [master](https://github.com/dallinger/dallinger/tree/master) (xxxx-xx-xx)
+- Initial docker support: `dallinger docker debug` command to run dallinger in a local container
 
 ## [v7.0.0](https://github.com/dallinger/dallinger/tree/v7.0.0) (2021-02-11)
 - Infrastructure: Drop support for python 2.x :-/